warehouse_agent_vertex/requirements.txt
pandas
sqlalchemy
pybigquery
google-cloud-bigquery
google-auth

google-cloud-aiplatform
langchain
langchain-community
langchain-google-vertexai

prophet
mlxtend

fastapi
uvicorn
python-dotenv

apache-airflow
apache-airflow-providers-google
